vcl: check for null device
The xmloff/draw test series was crashing soon after "testTextBoxLoss"
finished. Even though it finished successfully, CppUnitTest was crashing
when popping the protector. From the backtrace, the crash resulted from
attempts to access null pointers.

I do not know what caused those pointers to be null, but I added the
checks so that it could at least proceed to the other tests.

Change-Id: Ia0014a273142442a791a43fb9e3f75c2452bbd45
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143332
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>