tdf#141386 svx,sw: fix setting name/title/description on SwVirtFlyDrawObj
The problem is that setting the title via the dialog in Writer calls
SwFlyFrameFormat::SetObjTitle() which sets SwFrameFormat::msTitle and
SdrObject::m_pPlusData::aObjTitle on the master SwFlyDrawObj.

Now ViewContactOfSdrObj::embedToObjectSpecificInformation() is called
for the SwVirtFlyDrawObj, not the SwFlyDrawObj, because that one is not
on the SdrPage.

However, SwFlyFrameFormat::GetObjTitle() will return
SwFrameFormat::msTitle as a fallback.

So what's missing here is SwVirtFlyDrawObj must forward calls that
get/set model properties to its associated SwFlyDrawObj - implement that
in base class SdrVirtObj.

But note that this does not fix the reported PDF export bug - that was
already fixed by commit 8f1c340ebd51353a3a058d02e356d0c60c3e5849 in sw
and this fix is for other users of ObjectInfoPrimitive2D.

Change-Id: Id306c44c67c069777373e5e4d55415176b513afc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143465
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>