gb_LinkTarget_use_clang: only use flto=thin with clang, not with gcc
The code always used flto=thin, but gcc doesn't know that switch and
fails. The gb_Library_use_clang used by Skia is actually a prefer clang
option, and the corresponding gb_LinkTarget_use_clang macro accounts for
that by using "$(or clangvalue,gccvalue)" to define variables.

This is OK for stuff that defaults to empty value when no clang is found
on the system, but gb_CLANG_LTOFLAGS is defined unconditionally (in
com_GCC_defs.mk), so using the $(or..) function is completly bypassed,
always using the flto=thin switch.

Change-Id: I903e5e4818b8ecbe63ed85952b8741140972e8d7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143572
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>