sd: Don't replace table styles on page pasting
i.e. insert a table, and set its style to anything except
default. Then right click on the page in the page pane =>
Copy. Right click again => Paste. The table on the first
page will reset to the default table style.

The problem is in SdStyleSheetPool::CopyTableStyles which
attempts to replace existing styles with the pasted ones,
causing the old ones to be destroyed. But in fact there's
absolutely no case where we want to do anything like this.
All we want is to keep the formatting of the pasted page,
not to change existing styles which might be used in the
rest of the document. (And if the pasted style has indeed
a different formatting, it can be inserted under a
different name. To be handled in a follow-up commit.)

Change-Id: I7caaffc72265bb045bf21c831f2da27012e41845
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143644
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>