sw, .uno:InsertBookmark: add a new BookmarkText parameter and accept HTML there
There was already an UNO command to insert a new bookmark with the
provided name, in a non-interactive way.

What was missing is to allow specifying the bookmark text, which is to
some extent not part of the bookmark, but e.g. the bookmark dialog
allows editing that still.

Add a new BookmarkText parameter to .uno:InsertBookmark, in case it's
specified then we interpret this as HTML and we create the bookmark on
the imported content, not simply at the current cursor position.

This is similar to commit 5ea41bbca8efa7050c04bd572c212f61f13cd074 (sw,
.uno:TextFormField command: accept HTML in the FieldResult parameter,
2022-11-16), but that was for the field mode, while this is for the
bookmark mode of Zotero.

Change-Id: I4928d173e197796d40fdb53f81e84b6bfd77cdc5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143736
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins