sw layout: invalidate margins of body content when moving a fly from page
The (simplified) bugdoc has 9 pages, and once the ToC is updated, the
two tables below the (only) bookmark in the document has a 5879 twips
gap between them for no reason. Adding a new paragraph to the second
table "fixes" the incremental layout.

What seems to happen is that there is an anchored image on page 8, but
the layout decides that it doesn't fit there so it gets moved to page 9.
Before this move happens, SwTabFrame::Format() handles the outer of the
nested table below the unwanted gap and its CalcFlyOffsets() notices
that a large upper margin is wanted so that the table wraps around the
anchored image. Later the image gets moved to the next page, but the
removal doesn't invalidate the print area of the table with the large
top margin, so the unwanted gap appears.

Fix the problem by looking at the old page in SwPageFrame::MoveFly() and
walking through the immediate children of the body frame: if the frame
print area is invalidated, then SetYMargins() is called from
SwTabFrame::Format(), which also invalidates the size of the outer table
frame, so it will use Shrink() to have a correct size.

This appears to be a regression from commit
dd01b0b10380235061ff7700065c3496a7e27b5b (tdf#134298 sw: layout: remove
left-over page frame without content, 2020-11-13).

Change-Id: I0424d9eea4d8a030959f8534985950c7efad4686
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143881
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins