sw, createTextRangeByPixelPosition(): fix crash with at-page anchored images
When createTextRangeByPixelPosition() was called with a position that is
inside an at-page anchored image, we crashed.

The problem is that in general we can't create an XTextRange for a
graphic node. We already tried to return the anchor position for graphic
nodes since 0769d680bcc49e89b32476dec86f7111a00c41b8 (sw,
createTextRangeByPixelPosition(): fix crash when the position is an
image, 2022-11-22), but that fixed the problem only in case
GetContentAnchor() gave us a doc model position.

Fix the problem by returning an empty reference in the "graphic node but
no content anchor" case, that is still better for at-page and at-frame
anchored images than a crash.

A follow-up change could be to add a new API that has a more generic
return type (similar to how getSelection() returns `any`) and then
return the image itself, not its anchor in the image case. That approach
would work for at-page and at-frame anchored images as well.

Change-Id: If850ec945c24052568230a83da2ae2aa004b4265
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144043
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins