ScriptForge - (SF_Datasheet) default values in GetText() and GetValue()
Before this commit, the 'column' argument was manadatory.
Now the argument becomes optional and designates
the current column, i.e. the column containing the cursor.

Additionally, when a datasheet is closed by code,
any filter is better removed to avoid eventual
user misunderstandings when the datasheet is reopened
manually during the same session.

Change-Id: I904cfa56a9407539a89b7d2451dcf6b45292f660
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144242
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins