tdf#151548 vba ContentControls: fix fundamental error
The problem was that the SwTextControl could be deleted
by the user or in VBA and my references (or pointers
for that matter) to m_rCC would be invalid.

Instead, use the shared_ptr, which will always be valid,
from the SwContentControl. Thankfully, there is
backwards access to the Format and Text controls.

This is nicer anyways. It has reduced code lines,
and simpler access to the most used aspect.

Change-Id: Ie382b33ecda5c3d0024f0eabcb8ff158db942f52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144345
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>