sw: ODF import: fix assert in SwTable::ConvertSubtableBox()
Assert in swnewtable.cxx:2179 fires:

For ooo98643-1.odt the conversion is not possible, the default for row
height is "variable" which wasn't handled.

For ooo27428-1.sxw the conversion is possible so remove the assert;
the sum of the inner row fixed height is just smaller than the outer
row's but unsurprisingly the subtables is layouted such that the last
row gets the extra space, same as the conversion does.

(regression from commit e14afbea98b9b20030d78d2660ca761d796a56e9)

Change-Id: I2635179f350c8b5b84016dc82625b73364df92a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144622
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>