loplugin:unocast (sd::TextApiObject)
(See the upcoming commit introducing that loplugin:unocast on why such
dynamic_casts from UNO types are dangerous.)

The odd implementation of TextApiObject::getImplementation had been effectively
like that ever since the code's introduction in
5d431414afa79d392fe6a885940c559236c25aa2 "CWS-TOOLING: integrate CWS
impressnotes01", for no obvious reason.  And as it can be reduced to a single
comphelper::getFromUnoTunnel call, drop the function and directly call that at
the (single) call site.

Change-Id: If50a81cf8bee0dddbdfb5e6b5516c24a03e858e9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144699
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>