tdf#127588 - extend selection to the entire field if nothing is selected
If a user pastes a string to an input field that has a maximum length restriction and no text is selected in the input field, extend the selection to the respective length of the pasted string or the maximum text length of the input field in order to prevent a text truncation warning. If the user selects some text in the input field, just replace the currently selected text and show the text truncation warning as before.

Change-Id: I2115f8c6672c9c66770abfa6918cf327a7eece17
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144784
Tested-by: Jenkins
Reviewed-by: Andreas Heinisch <andreas.heinisch@yahoo.de>