tdf#151548 DOCX formfield: export checkbox name
The two other fields are exporting the name,
(although it is getting somewhat lost in excess bookmarks),
because they are just using the Fieldmark name property.

See LO 7.5 eb9828a8623f2cee82b7a31e929e016b516af2f7
for tdf#151548 sw: use provided name for formfields
which saved the name for checkboxes and dropdowns.

ODF_FORMCHECKBOX_NAME was simply used as a grabbag
for ww8import. Since that is obsolete, get rid of it.

mostly introduced with ---
commit b547c0c2aa901667fef85233282ec84f34b7e5f4
Author: Noel Power on Thu Feb 10 16:18:40 2011 +0000
    some form field import/export improvements

ODF_FORMDROPDOWN_NAME was dropped early on. ---
commit 79770ff55bd8c3bc5948c51373e8cb7867ce43d3
Author: Noel Power on Tue Mar 1 12:30:23 2011 +0000
    partial revert of 803409125f4ed896b391acb99266d50691c6fd4a

    remove currently unnecessary ( but possibly future )
    odf attributes for form field controls to prevent those
    attributes from being written to odf

Change-Id: I133028dce65361314e663bb5238f99c607f0ab14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144792
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>