tdf#56258 tdf#106988 sw numbering IVX: fix bad indent/tab value
The comment (and the previous code) indicates a progressive
increment by 1.3cm for each level. Well, the third level
was some mid-way-between-level1-and-level2 anomaly.
So just drop it and add the proper indent for the last level.

The tenth level should be 10 times larger than the first indent.
Previously it was 283 and 2835, so the current values
of 754 and 7541 sound perfect.

This was a LO 6.0 regression by
commit 09ac7500ad9038e5a8ec2493f338d25453f61e65

Change-Id: I41623424ab5c768bdc3914ae9b95d02631bb1a2e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144827
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>