SwNavigator: Improve mouse move handler for bringing objects to
attention in the document view

Fixes crash that happens in SalInstanceTreeView iter_compare when the
mouse enters the content tree at a point that has no tree entry and
is then moved over an entry.

Moves code to bring entry content to attention in the document view
from the mouse move handler to it's own function.

Renames the overlay timer handler function and moves it to the area
of the code having related functions.

Checks added for visible content and if the type has any content to
bring to attention.

Change-Id: I381eed6ffcf7ddb095c4c09650578a7f7fd874b0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145168
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>